+++ CopilotKit/packages/react-core/src/hooks/use-chat.ts
@@ -97,6 +97,31 @@ export type UseChatOptions = {
    * setState-powered method to update the agent session
    */
   setAgentSession: React.Dispatch<React.SetStateAction<AgentSession | null>>;
+
+  /**
+   * The current thread ID.
+   */
+  threadId: string | null;
+
+  /**
+   * set the current thread ID
+   */
+  setThreadId: (threadId: string | null) => void;
+
+  /**
+   * The current run ID.
+   */
+  runId: string | null;
+
+  /**
+   * set the current run ID
+   */
+  setRunId: (runId: string | null) => void;
+
+  /**
+   * The global chat abort controller.
+   */
+  chatAbortControllerRef: React.MutableRefObject<AbortController | null>;
 };
 
 export type UseChatHelpers = {
@@ -134,11 +159,12 @@ export function useChat(options: UseChatOptions): UseChatHelpers {
     coagentStates,
     agentSession,
     setAgentSession,
+    threadId,
+    setThreadId,
+    runId,
+    setRunId,
+    chatAbortControllerRef,
   } = options;
-
-  const abortControllerRef = useRef<AbortController>();
-  const threadIdRef = useRef<string | null>(null);
-  const runIdRef = useRef<string | null>(null);
 
   const runChatCompletionRef = useRef<(previousMessages: Message[]) => Promise<Message[]>>();
   // We need to keep a ref of coagent states because of renderAndWait - making sure
@@ -148,6 +174,10 @@ export function useChat(options: UseChatOptions): UseChatHelpers {
   coagentStatesRef.current = coagentStates;
   const agentSessionRef = useRef<AgentSession | null>(agentSession);
   agentSessionRef.current = agentSession;
+  const threadIdRef = useRef<string | null>(threadId);
+  threadIdRef.current = threadId;
+  const runIdRef = useRef<string | null>(runId);
+  runIdRef.current = runId;
 
   const publicApiKey = copilotConfig.publicApiKey;
 
@@ -175,7 +205,7 @@ export function useChat(options: UseChatOptions): UseChatHelpers {
       }),
     ];
     const abortController = new AbortController();
-    abortControllerRef.current = abortController;
+    chatAbortControllerRef.current = abortController;
 
     setMessages([...previousMessages, ...newMessages]);
 
@@ -231,7 +261,7 @@ export function useChat(options: UseChatOptions): UseChatHelpers {
           })),
         },
         properties: copilotConfig.properties,
-        signal: abortControllerRef.current?.signal,
+        signal: chatAbortControllerRef.current?.signal,
       }),
     );
 
@@ -258,6 +288,9 @@ export function useChat(options: UseChatOptions): UseChatHelpers {
 
         threadIdRef.current = value.generateCopilotResponse.threadId || null;
         runIdRef.current = value.generateCopilotResponse.runId || null;
+
+        setThreadId(threadIdRef.current);
+        setRunId(runIdRef.current);
 
         const messages = convertGqlOutputToMessages(
           filterAdjacentAgentStateMessages(value.generateCopilotResponse.messages),
@@ -442,7 +475,7 @@ export function useChat(options: UseChatOptions): UseChatHelpers {
   };
 
   const stop = (): void => {
-    abortControllerRef.current?.abort();
+    chatAbortControllerRef.current?.abort();
   };
 
   return {